spun off input & output
